+++ MiDAS/R/MiDAS_utils.R
@@ -184,7 +184,7 @@ getVariableAAPos <- function(alignment,
 #'   for expression values numeric vector will be returned.
 #'
 #' @examples
-#' dictionary <- system.file("extdata", "Match_4digit_supertype.txt", package = "MiDAS")
+#' dictionary <- system.file("extdata", "Match_allele_HLA_supertype.txt", package = "MiDAS")
 #' convertAlleleToVariable(c("A*01:01", "A*02:01"), dictionary = dictionary)
 #'
 #' @importFrom assertthat assert_that is.string is.readable see_if
@@ -208,6 +208,7 @@ convertAlleleToVariable <- function(allele,
     dictionary <- read.table(
       file = dictionary,
       header = TRUE,
+      sep = "\t",
       stringsAsFactors = FALSE
     )
   }
@@ -613,14 +614,18 @@ assertthat::on_failure(isFlagOrNULL) <- function(call, env) {
 #'
 #' @param file.names Logical value. If FALSE, only the names of dictionaries are
 #' returned. If TRUE their file names are returned.
+#' @param pattern String used to match dictionary names, it can be a regular
+#'   expression.
 #'
 #' @return Character vector with names of available HLA alleles dictionaries.
 #'
 #' @export
-listMiDASDictionaries <- function(file.names = FALSE) {
+listMiDASDictionaries <- function(pattern = ".*",
+                                  file.names = FALSE) {
+  pattern <- paste0("^Match.*", pattern, ".*.txt$")
   lib <- list.files(
     path = system.file("extdata", package = "MiDAS"),
-    pattern = "^Match_.*.txt$",
+    pattern = pattern,
     full.names = file.names
   )
 
@@ -684,7 +689,7 @@ assertthat::on_failure(characterMatches) <- function(call, env) {
 #'   in \code{x} are removed during conversion.
 #'
 #' @examples
-#' x <- c("1+3|16+3", "1+1", NA)
+#' x <- c(NA, "1+3|16+3", "1+1", NA)
 #' kirHaplotypeToCounts(x)
 #'
 #' @importFrom assertthat assert_that is.flag is.readable see_if
@@ -693,16 +698,16 @@ assertthat::on_failure(characterMatches) <- function(call, env) {
 #'
 #' @export
 kirHaplotypeToCounts <- function(x,
-                                 hap_dict = system.file("extdata", "Match_KIR_haplotype_genes.tsv", package = "MiDAS"),
+                                 hap_dict = system.file("extdata", "Match_kir_haplotype_gene.txt", package = "MiDAS"),
                                  binary = TRUE) {
   assert_that(
     is.character(x),
     is.readable(hap_dict),
     is.flag(binary)
   )
-  hap_dict <- read.table(hap_dict)
+  hap_dict <- read.table(hap_dict, stringsAsFactors = FALSE)
 
-  x <- na.omit(x)
+  # x <- na.omit(x)
   x_split <- stri_split_fixed(x, "|")
   x_split_unlist <- unlist(x_split)
 
@@ -710,7 +715,7 @@ kirHaplotypeToCounts <- function(x,
   assert_that(
     all(haps_match <- vapply(
       X = x_split_unlist_haps,
-      FUN = function(x) all(x %in% rownames(hap_dict)),
+      FUN = function(x) all(na.omit(x) %in% rownames(hap_dict)),
       FUN.VALUE = logical(1)
     )),
     msg = sprintf(
@@ -730,10 +735,11 @@ kirHaplotypeToCounts <- function(x,
   counts <- vapply(
     X = x_split,
     FUN = function(hap) {
+      hap <- ifelse(is.na(hap), NA, hap) # class character NA cannot be used for columns indexing
       hap <- counts[, hap, drop = FALSE]
       if (ncol(hap) > 1) {
         assert_that(
-          all(apply(hap, 2, function(col) all(col == hap[, 1]))),
+          all(apply(hap, 2, function(col) all(col == hap[, 1])), na.rm = TRUE),
           msg = sprintf("haplotype %s can not be unambigously converted to counts", hap)
         )
       }
@@ -744,7 +750,7 @@ kirHaplotypeToCounts <- function(x,
   )
   counts <- t(counts)
   counts <- as.data.frame(counts, optional = TRUE, stringsAsFactors = FALSE)
-  counts <- cbind(haplotypes = x, counts)
+  counts <- cbind(haplotypes = x, counts, stringsAsFactors = FALSE)
 
   return(counts)
 }